Move utility from `base` to `protobuf`.
