+++ base/src/test/java/io/spine/protobuf/FieldPathsTest.java
@@ -18,23 +18,25 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.base;
+package io.spine.protobuf;
 
 import com.google.common.testing.NullPointerTester;
 import com.google.protobuf.Any;
 import com.google.protobuf.Empty;
-import io.spine.test.base.AnyHolder;
-import io.spine.test.base.GenericHolder;
-import io.spine.test.base.StringHolder;
-import io.spine.test.base.StringHolderHolder;
+import io.spine.base.FieldPath;
+import io.spine.base.Time;
+import io.spine.test.protobuf.AnyHolder;
+import io.spine.test.protobuf.GenericHolder;
+import io.spine.test.protobuf.StringHolder;
+import io.spine.test.protobuf.StringHolderHolder;
 import io.spine.testing.UtilityClassTest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import static com.google.common.truth.Truth.assertThat;
-import static io.spine.base.FieldPaths.fieldAt;
-import static io.spine.base.FieldPaths.parse;
 import static io.spine.protobuf.AnyPacker.pack;
+import static io.spine.protobuf.FieldPaths.fieldAt;
+import static io.spine.protobuf.FieldPaths.parse;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
